+++ doc/source/changelog.rst
@@ -9,7 +9,11 @@ Changelog
 ------------------
 .. note:: This version is not yet released and is under active development.
 
-* ...
+* Add new meths:
+
+  - :py:meth:`~AbstractSerialStream.discard_input`
+  - :py:meth:`~AbstractSerialStream.discard_output`
+  - :py:meth:`~AbstractSerialStream.send_break`
 
 .. _changelog.0.1.0: